Correct rendering of pipeline in steps
This change forced us to alter the pipeline-card into a new view.  What
we were aiming for was apparently not a card, but a collection view in
this case.  The card still exists, but it now actually behaves like a
card.  Which is probably what you'de expect, given the terminology ;-)