Add Dockerfile for easy hosting
We are downloading the version from GitHub.  For this, the last version
needs to be uploaded as a dist.zip file on the last release.  With this,
we don't require changes to happen on the Dockerfile when a new release
is published, but a manual build will likely need to be scheduled.